py29/cinema/requirements.txt
django 
djangorestframework
djangorestframework-simplejwt
django-filter
psycopg2-binary
drf-yasg
python-decouple
pillow
